requirements.txt
alembic==1.0.10
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
colorama==0.4.1
Flask==1.0.3
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.0
importlib-metadata==0.17
itsdangerous==1.1.0
Jinja2==2.10.1
Mako==1.0.12
MarkupSafe==1.1.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pyparsing==2.4.0
pytest==4.6.2
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
six==1.12.0
SQLAlchemy==1.3.4
wcwidth==0.1.7
Werkzeug==0.15.4
zipp==0.5.1
